[HER-11]Follow-up-Move const to new file
