+++ src/home/carousel/hooks/use-on-arrow-click.js
@@ -1,15 +1,6 @@
 import { useContext } from 'home/carousel/hooks/use-context'
-import { slides } from 'home/carousel/components'
 import { useActions } from 'home/carousel/hooks/use-actions'
-
-const calculateParameters = (direction, index) => {
-  const numSlides = slides.length
-  const increment = direction === 'left' ? -1 : 1
-  const newIndex = (index + increment + numSlides) % numSlides
-  const oppDirection = direction === 'left' ? 'right' : 'left'
-
-  return { newIndex, oppDirection }
-}
+import { calculateParameters } from 'home/carousel/caculate-parameters'
 
 export const useOnArrowClick = () => {
   const { index } = useContext()
